# contracts/public-openapi.yaml
openapi: 3.1.0
info:
  title: Barq Server API
  version: 0.3.0
  description: Public tenant data and webhook control API for BarqDB.
servers:
  - url: /
security:
  - serviceKey: []
paths:
...
  /v1/admin/tenants:
    get:
      summary: List tenants visible to the admin key
      operationId: listTenants
      responses:
        "200":
          description: Tenant registry.
          content:
            application/json:
              schema:
                type: object
                required: [tenants]
                properties: {tenants: {type: array, items: {$ref: "#/components/schemas/Tenant"}}}
    post:
      summary: Register a tenant
      description: Requires a global admin key. This registers logical databases but never deletes or rewrites Barq data.
      operationId: createTenant
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateTenant"}
      responses: {"201": {description: Tenant registered.}, "409": {description: Tenant already exists.}}
  /v1/admin/tenants/{id}:
    parameters:
      - {$ref: "#/components/parameters/id"}
    get:
      summary: Get a tenant
      operationId: getTenant
      responses: {"200": {description: Tenant record.}}
    patch:
      summary: Update or enable a tenant
      description: Requires a global admin key. Removing a database stops its scoped keys and polling, but leaves its data untouched.
      operationId: updateTenant
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateTenant"}
      responses: {"200": {description: Tenant updated.}}
    delete:
      summary: Disable a tenant without deleting data
      operationId: disableTenant
      responses: {"204": {description: Tenant disabled.}}
  /v1/admin/api-keys:
    get:
      summary: List API key metadata in the caller scope
      description: Raw keys and digests are never returned.
      operationId: listAPIKeys
      responses:
        "200":
          description: API key metadata.
          content:
            application/json:
              schema:
                type: object
                required: [api_keys]
                properties: {api_keys: {type: array, items: {$ref: "#/components/schemas/ServiceKey"}}}
    post:
      summary: Create an API key
      description: The raw key is returned once and only its SHA-256 digest is stored in control.barq.
      operationId: createAPIKey
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateServiceKey"}
      responses:
        "201":
          description: API key metadata and one-time secret.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CreatedServiceKey"}
  /v1/admin/api-keys/{id}:
    parameters:
      - {$ref: "#/components/parameters/id"}
    get:
      summary: Get API key metadata
      operationId: getAPIKey
      responses: {"200": {description: API key metadata.}}
    patch:
      summary: Update API key label, actions, or enabled state
      operationId: updateAPIKey
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateServiceKey"}
      responses: {"200": {description: API key updated.}, "409": {description: The last global admin key cannot be disabled.}}
    delete:
      summary: Revoke an API key but retain audit metadata
      operationId: revokeAPIKey
      responses: {"204": {description: API key revoked.}, "409": {description: The last global admin key cannot be revoked.}}
...
  /v1/tenants/{tenant}/databases/{database}/objects/{type}/{primaryKey}:
    get:
      summary: Read an object
      operationId: readObject
      parameters:
        - {$ref: "#/components/parameters/tenant"}
        - {$ref: "#/components/parameters/database"}
        - {$ref: "#/components/parameters/type"}
        - {$ref: "#/components/parameters/primaryKey"}
      responses: {"200": {description: Object and ETag.}, "404": {description: Object not found.}}
    patch:
      summary: Patch an object
      operationId: patchObject
      parameters:
        - {$ref: "#/components/parameters/tenant"}
        - {$ref: "#/components/parameters/database"}
        - {$ref: "#/components/parameters/type"}
        - {$ref: "#/components/parameters/primaryKey"}
      responses: {"200": {description: Object updated.}, "409": {description: Stale ETag.}}
    delete:
      summary: Delete an object
      operationId: deleteObject
      parameters:
        - {$ref: "#/components/parameters/tenant"}
        - {$ref: "#/components/parameters/database"}
        - {$ref: "#/components/parameters/type"}
        - {$ref: "#/components/parameters/primaryKey"}
      responses: {"200": {description: Object deleted.}, "409": {description: Stale ETag.}}
  /v1/tenants/{tenant}/databases/{database}/objects/{type}:query:
    post:
      summary: Query objects
      operationId: queryObjects
      parameters:
        - {$ref: "#/components/parameters/tenant"}
        - {$ref: "#/components/parameters/database"}
        - {$ref: "#/components/parameters/type"}
      responses: {"200": {description: Cursor-paged query result.}}
  /v1/tenants/{tenant}/databases/{database}:batch:
    post:
      summary: Commit an atomic batch
      operationId: executeBatch
      parameters:
        - {$ref: "#/components/parameters/tenant"}
        - {$ref: "#/components/parameters/database"}
      responses: {"200": {description: Atomic batch result.}}
...
  /v1/webhooks:
    get:
      summary: List webhooks
      operationId: listWebhooks
      responses: {"200": {description: Visible webhook registrations.}}
    post:
      summary: Register a webhook
      operationId: registerWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/WebhookRegistration"}
      responses:
        "201":
          description: Webhook and one-time signing secret.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RegisteredWebhook"}
  /v1/webhooks/{id}:
    parameters:
      - {$ref: "#/components/parameters/id"}
    get:
      summary: Get a webhook
      operationId: getWebhook
      responses: {"200": {description: Webhook registration.}}
    patch:
      summary: Update a webhook
      operationId: updateWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/WebhookRegistration"}
      responses: {"200": {description: New immutable revision activated.}}
    delete:
      summary: Disable a webhook
      operationId: disableWebhook
      responses: {"204": {description: Webhook disabled.}}
...
components:
  securitySchemes:
    serviceKey: {type: http, scheme: bearer}
  parameters:
    tenant: {name: tenant, in: path, required: true, schema: {type: string}}
    database: {name: database, in: path, required: true, schema: {type: string}}
    type: {name: type, in: path, required: true, schema: {type: string}}
    primaryKey: {name: primaryKey, in: path, required: true, schema: {type: string}}
    id: {name: id, in: path, required: true, schema: {type: string}}
  schemas:
    Tenant:
      type: object
      additionalProperties: false
      required: [id, name, databases, enabled, created_at, updated_at]
      properties:
        id: {type: string, pattern: "^[A-Za-z0-9][A-Za-z0-9_-]{0,62}$"}
        name: {type: string, minLength: 1, maxLength: 120}
        databases: {type: array, minItems: 1, maxItems: 50, uniqueItems: true, items: {type: string}}
        enabled: {type: boolean}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    CreateTenant:
      type: object
      additionalProperties: false
      required: [id, name, databases]
      properties:
        id: {type: string, pattern: "^[A-Za-z0-9][A-Za-z0-9_-]{0,62}$"}
        name: {type: string, minLength: 1, maxLength: 120}
        databases: {type: array, minItems: 1, maxItems: 50, uniqueItems: true, items: {type: string}}
    UpdateTenant:
      type: object
      additionalProperties: false
      properties:
        name: {type: string, minLength: 1, maxLength: 120}
        databases: {type: array, minItems: 1, maxItems: 50, uniqueItems: true, items: {type: string}}
        enabled: {type: boolean}
    ServiceKey:
      type: object
      additionalProperties: false
      required: [id, tenant, database, actions, enabled, created_at, updated_at]
      properties:
        id: {type: string}
        label: {type: string, maxLength: 120}
        tenant: {type: string}
        database: {type: string}
        actions: {type: array, minItems: 1, maxItems: 20, uniqueItems: true, items: {type: string}}
        enabled: {type: boolean}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    CreateServiceKey:
      type: object
      additionalProperties: false
      required: [label, tenant, database, actions]
      properties:
        label: {type: string, minLength: 1, maxLength: 120}
        tenant: {type: string}
        database: {type: string}
        actions: {type: array, minItems: 1, maxItems: 20, uniqueItems: true, items: {type: string}}
    UpdateServiceKey:
      type: object
      additionalProperties: false
      properties:
        label: {type: string, maxLength: 120}
        actions: {type: array, minItems: 1, maxItems: 20, uniqueItems: true, items: {type: string}}
        enabled: {type: boolean}
    CreatedServiceKey:
      type: object
      additionalProperties: false
      required: [key, secret]
      properties:
        key: {$ref: "#/components/schemas/ServiceKey"}
        secret: {type: string, description: Returned once and never stored.}
...
    Scope:
      type: object
      additionalProperties: false
      required: [tenant, database]
      properties:
        tenant: {type: string}
        database: {type: string}
    Transform:
      type: object
      additionalProperties: false
      required: [language, source]
      properties:
        language: {type: string, const: javascript}
        source: {type: string, description: Sandboxed JavaScript defining transform(ctx) and optional filter(ctx).}
    WebhookRegistration:
      type: object
      additionalProperties: false
      required: [name, scope, url, events, transform]
      properties:
        name: {type: string}
        scope: {$ref: "#/components/schemas/Scope"}
        url: {type: string, format: uri}
        events: {type: array, minItems: 1, items: {type: string}}
        object_types: {type: array, items: {type: string}}
        reads: {type: array, maxItems: 5, items: {type: object}}
        transform: {$ref: "#/components/schemas/Transform"}
        limits: {type: object}
    Webhook:
      type: object
      required: [id, name, scope, url, events, active_revision, enabled]
      properties:
        id: {type: string}
        name: {type: string}
        scope: {$ref: "#/components/schemas/Scope"}
        url: {type: string, format: uri}
        events: {type: array, items: {type: string}}
        object_types: {type: array, items: {type: string}}
        active_revision: {type: integer, minimum: 1}
        enabled: {type: boolean}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    RegisteredWebhook:
      type: object
      required: [webhook, secret]
      properties:
        webhook: {$ref: "#/components/schemas/Webhook"}
        secret: {type: string, description: Returned once. Store it securely.}
...